# shiftory-server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Shiftory API
  version: 1.0.0
  description: 排班协同、导入预览与团队日历 API。业务接口统一返回 `{data, requestId}`，错误返回 `{error, requestId}`。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /workspaces/{workspaceId}/shifts:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [Shifts]
      operationId: listShifts
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Shifts]
      operationId: createShift
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ShiftInput' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
...
  /workspaces/{workspaceId}/schedules/{userId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Schedules]
      operationId: listSchedules
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /workspaces/{workspaceId}/schedules/{userId}/{date}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - $ref: '#/components/parameters/UserId'
      - $ref: '#/components/parameters/Date'
    put:
      tags: [Schedules]
      operationId: upsertSchedule
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
...
    Date:
      name: date
      in: path
      required: true
      schema: { type: string, format: date }
    StartDate:
      name: start
      in: query
      required: true
      schema: { type: string, format: date }
    EndDate:
      name: end
      in: query
      required: true
      schema: { type: string, format: date }
...
  responses:
    Success:
      description: 成功
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SuccessEnvelope' }
...
    Error:
      description: 失败
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    SuccessEnvelope:
      type: object
      required: [data, requestId]
      properties:
        data: {}
        requestId: { type: string, format: uuid }
    ErrorEnvelope:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
        requestId: { type: string, format: uuid }
...
    ShiftInput:
      type: object
      required: [name, code, crossDay, displayColor]
      properties:
        name: { type: string, maxLength: 80 }
        code: { type: string, maxLength: 64 }
        startTime: { type: [string, 'null'], pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        endTime: { type: [string, 'null'], pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        crossDay: { type: boolean }
        displayColor: { type: string }
        enabled: { type: boolean, default: true }
        sortOrder: { type: integer, default: 0 }
        aliases:
          type: array
          items: { type: string, maxLength: 80 }
    ScheduleSegment:
      type: object
      required: [type, crossDay]
      properties:
        type: { type: string, enum: [SHIFT, TIME_RANGE] }
        shiftId: { type: integer, format: int64 }
        startTime: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        endTime: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        crossDay: { type: boolean }
    ScheduleInput:
      type: object
      required: [status, segments]
      properties:
        status: { type: string, enum: [WORKING, REST] }
        note: { type: string, maxLength: 1000 }
        version: { type: integer, format: int64, minimum: 0 }
        segments:
          type: array
          maxItems: 8
          items: { $ref: '#/components/schemas/ScheduleSegment' }
...